backend/requirements.txt
fastapi
mangum
openai
python-dotenv
pydantic
uvicorn[standard]